@@ -15,8 +15,8 @@
     "moment": "^2.22.2",
     "node-sass": "^4.9.0",
     "npm-run-all": "^4.1.3",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-icons-weather": "github:milesdelliott/react-weather-icons#feature/night-icons",
     "react-scripts": "1.1.4",
     "save": "^2.3.2",
@@ -30,6 +30,7 @@
     "build-js": "react-scripts build",
     "build": "npm-run-all build-css build-js",
     "test": "react-scripts test --env=jsdom",
+    "test:debug": "react-scripts --inspect-brk test --runInBand --no-cache",
     "eject": "react-scripts eject"
   },
   "devDependencies": {
